+++ winit-core/src/lib.rs
@@ -13,5 +13,6 @@ compile_error! { "no-std and no-alloc usage are not yet supported" }
 pub mod dpi;
 pub mod error;
 pub mod event;
+pub mod event_loop;
 pub mod keyboard;
 pub mod window;
