+++ winit-core/src/window.rs
@@ -191,3 +191,25 @@ impl Default for ImePurpose {
         Self::Normal
     }
 }
+
+/// An opaque token used to activate the [`Window`].
+///
+/// [`Window`]: crate::window::Window
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct ActivationToken {
+    pub(crate) token: String,
+}
+
+impl ActivationToken {
+    pub fn new(_token: String) -> Self {
+        Self { token: _token }
+    }
+
+    pub fn token(&self) -> &str {
+        &self.token
+    }
+
+    pub fn into_token(self) -> String {
+        self.token
+    }
+}
