+++ winit/src/platform_impl/linux/wayland/event_loop/sink.rs
@@ -26,10 +26,8 @@ impl EventSink {
     /// Add new device event to a queue.
     #[inline]
     pub fn push_device_event(&mut self, event: DeviceEvent, device_id: DeviceId) {
-        self.window_events.push(Event::DeviceEvent {
-            event,
-            device_id,
-        });
+        self.window_events
+            .push(Event::DeviceEvent { event, device_id });
     }
 
     /// Add new window event to a queue.
