+++ winit/src/platform_impl/linux/x11/window.rs
@@ -588,7 +588,7 @@ impl UnownedWindow {
 
         // Remove the startup notification if we have one.
         if let Some(startup) = window_attrs.platform_specific.activation_token.as_ref() {
-            leap!(xconn.remove_activation_token(xwindow, &startup._token));
+            leap!(xconn.remove_activation_token(xwindow, startup.token()));
         }
 
         // We never want to give the user a broken window, since by then, it's too late to handle.
