+++ winit/src/window.rs
@@ -1596,17 +1596,3 @@ pub enum Fullscreen {
     /// Providing `None` to `Borderless` will fullscreen on the current monitor.
     Borderless(Option<MonitorHandle>),
 }
-
-/// An opaque token used to activate the [`Window`].
-///
-/// [`Window`]: crate::window::Window
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct ActivationToken {
-    pub(crate) _token: String,
-}
-
-impl ActivationToken {
-    pub(crate) fn _new(_token: String) -> Self {
-        Self { _token }
-    }
-}
